// src/types/initiateMfaSetupMethod.ts
/**
 * Generated by orval v8.5.0 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */

export type InitiateMfaSetupMethod = typeof InitiateMfaSetupMethod[keyof typeof InitiateMfaSetupMethod];


export const InitiateMfaSetupMethod = {
  TOTP: 'TOTP',
  EMAIL: 'EMAIL',
  WEBAUTHN: 'WEBAUTHN',
} as const;